requirements.txt
amqp==2.5.0
billiard==3.6.0.0
certifi==2019.6.16
chardet==3.0.4
Django==1.11
django-background-tasks==1.2.0
django-compat==1.0.15
idna==2.8
kombu==4.6.3
mysql-connector-python==8.0.17
protobuf==3.9.1
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
vine==1.3.0
